pyproject.toml
[tool.poetry]
name = "repopoppin"
version = "0.1.0"
description = ""
authors = ["Hyeonu Kim <woo813625@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mysqlclient = "^2.2.3"
django = "^5.0.1"
djangorestframework = "^3.14.0"
pyjwt = "^2.8.0"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
